requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
pymongo==4.6.2
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
gunicorn==21.2.0
dnspython==2.6.1
requests==2.31.0
Pillow==10.2.0
python-dateutil==2.8.2